Introduce AbstractIntegrationTestWithWireMock and update IDs
Refactor integration tests to use AbstractIntegrationTestWithWireMock for consistent setup. Change data types in DTOs and entities from Byte to Integer for outputId and relayPort to ensure compatibility and correct handling of larger values.

+++ src/main/java/cz/patyk/solarmaxx/backend/client/ShellyProClient.java
@@ -14,9 +14,9 @@ import java.net.URI;
 public interface ShellyProClient {
     @Headers("Content-Type: application/json")
     @GetMapping("/rpc/Switch.GetStatus?id={outputId}")
-    ShellyProStatusOutputDto getOutputStatusWithSpecificPortObject(URI baseUrl, @PathVariable("outputId") byte outputId);
+    ShellyProStatusOutputDto getOutputStatusWithSpecificPortObject(URI baseUrl, @PathVariable("outputId") int outputId);
 
     @Headers("Content-Type: application/json")
     @GetMapping("/rpc/Switch.Toggle?id={outputId}")
-    ShellyProToggleOutputDto setOutputState(URI baseUrl, @PathVariable("outputId") byte outputId);
+    ShellyProToggleOutputDto setOutputState(URI baseUrl, @PathVariable("outputId") int outputId);
 }
